requirements.txt
fastapi>=0.108.0
uvicorn>=0.25.0
python-multipart>=0.0.6
pdf2image>=1.16.3
pytesseract>=0.3.10
langdetect>=1.0.9
markdown>=3.5.1
pydantic>=2.5.3
python-dotenv>=1.0.0
pytest>=7.4.3
pytest-cov>=4.1.0
httpx>=0.26.0
pytest-asyncio>=0.23.2
black>=23.12.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0
bandit>=1.7.6
safety>=2.3.5
hypothesis>=6.92.1
pytest-benchmark>=4.0.0
pillow>=10.1.0
aiofiles>=23.2.1
pytest-xdist>=3.5.0
opencv-python>=4.8.1.78
numpy>=1.26.2
xmltodict>=0.13.0
